+++ modules/index.js
@@ -23,6 +23,7 @@ export PropTypes from './PropTypes'
 export match from './match'
 export useRouterHistory from './useRouterHistory'
 export { formatPattern } from './PatternUtils'
+export applyRouterMiddleware from './applyRouterMiddleware'
 
 /* histories */
 export browserHistory from './browserHistory'